pyproject.toml
[project]
name = "youtube-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "elevenlabs>=1.51.0",
    "fal-client>=0.5.9",
    "firecrawl-py>=1.12.0",
    "google-api-python-client>=2.161.0",
    "google-cloud-texttospeech>=2.25.0",
    "google-genai>=1.8.0",
    "groq>=0.18.0",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.37",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.2.4",
    "langgraph>=0.2.74",
    "matplotlib>=3.10.0",
    "moviepy==1.0.3",
    "numpy>=1.26.4",
    "pillow==9.5.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tavily-python>=0.5.1",
]
